+++ frontend/src/types/Config.ts
