+++ frontend/src/types/CreateSessionReq.ts
@@ -1,4 +1,4 @@
-import { Config } from "./domain/config";
+import { Config } from "./Config";
 
 export type CreateSessionRes = {
   session_id: string;
